refactor: userDataRepository refactor
