+++ src/app/di/AppDiContainer.ts
@@ -1,7 +1,8 @@
 import { UserRepositoryImpl } from "../modules/Auth/data/user_credentials/UserRepositoryImpl";
 import { UserRepository } from "../modules/Auth/domain/user_credentials/UserRepository";
-import { UserDataRepositoryImpl } from "../modules/profile/data/UserDataRepositoryImpl";
-import { UserDataRepository } from "../modules/profile/domain/UserDataRepository";
+import { UserDataRepositoryImpl } from "../modules/repositories/user_data_repository/data/UserDataRepositoryImpl";
+import { UserDataRepository } from "../modules/repositories/user_data_repository/domain/UserDataRepository";
+
 import { DiContainer } from "./DiContainer";
 
 export class AppDiContainer implements DiContainer {
