+++ src/app/di/DiContainer.ts
@@ -1,5 +1,5 @@
 import { UserRepository } from "../modules/Auth/domain/user_credentials/UserRepository";
-import { UserDataRepository } from "../modules/profile/domain/UserDataRepository";
+import { UserDataRepository } from "../modules/repositories/user_data_repository/domain/UserDataRepository";
 
 export interface DiContainer {
   getUserRepository(): UserRepository;
