+++ src/app/modules/repositories/user_data_repository/data/TokenEntity.ts
@@ -1,0 +1,3 @@
+export interface TokenEntity {
+  token: string;
+}
