+++ src/app/modules/repositories/user_data_repository/domain/Token.ts
@@ -1,0 +1,3 @@
+export interface Token {
+  token: string;
+}
