+++ src/app/modules/repositories/user_data_repository/domain/UserDataRepository.ts
@@ -1,0 +1,5 @@
+import { Token } from "./Token";
+
+export interface UserDataRepository {
+  getUserData(): Promise<Token>;
+}
